Lowers volume of background in samples
Training ResNet20 and ResNet56 with background volumes uniformly
distributed between 0.1 and 0.4 of the background noise added to
recordings show models quickly converging to ~60% accuracy in training
and validation data sets. The loss also quickly converges as well.
Lowering the background volume under the hypothesis the the signal has
become too noisy with the previous settings.